1. BIGINT -> DECIMAL 2. RENAME sqlalchemy_starrocks TO sqlalchemy-starrocks

+++ contrib/sqlalchemy-connector/sqlalchemy_starrocks/dialect.py
@@ -1,6 +1,6 @@
 #! /usr/bin/python3
 # This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Inc.
-from sqlalchemy.dialects.mysql import BIGINT
+from sqlalchemy.dialects.mysql import DECIMAL
 from sqlalchemy.dialects.mysql.mysqldb import MySQLDialect_mysqldb
 
 class StarRocksDialect(MySQLDialect_mysqldb):
@@ -10,7 +10,7 @@ class StarRocksDialect(MySQLDialect_mysqldb):
     def __init__(self, *args, **kw):
         super(StarRocksDialect, self).__init__(*args, **kw)
         self.ischema_names = self.ischema_names.copy()
-        self.ischema_names['largeint']=BIGINT
+        self.ischema_names['largeint']=DECIMAL
 
     @classmethod
     def dbapi(cls):
